# Api/openapi.yaml
...
openapi: 3.0.3
info:
  title: servermander-api
  version: 2.0.0
  description: |
    Authenticated control-plane API for the box. Each SERVICE lives under its own
    path namespace - today that is DayZ (/dayz/*: allowlisted server actions plus
    read-only observability like log tail and host stats), with more to follow.
    Generic/host endpoints (/healthz, /sysload) and key management (/keys/*) sit at
    the root, outside any namespace. See README.md for the
    security model - the short version: every /dayz command is HMAC-signed over
    the exact raw body (wizard secret or a namespace-scoped derived key),
    destructive actions are player-guarded and warn players in-game before acting,
    and the service's entire root capability is one sudo-gated script with a
    closed verb set.
servers:
  - url: https://api.cytonicmushroom.ddns.net
    description: production (nginx TLS -> localhost Node)
  - url: http://127.0.0.1:3100
    description: on-box direct (bypasses nginx)
components:
  securitySchemes:
    HmacSignature:
      type: apiKey
      in: header
      name: X-Signature-256
      description: |
        `sha256=<hex>` where <hex> = HMAC-SHA256 over the EXACT raw request body
        bytes (empty string for bodyless requests). Keyed with the WIZARD secret
        (HMAC_SECRET in /etc/api/secrets.env on the box), or - when an X-Key-Id
        header names a derived key minted via /keys/create - that key's secret.
        A derived key is (capability x namespaces): it must hold the namespace of
        the path it calls (e.g. 'dayz' for /dayz/*, or '*'), else 403
        forbidden_namespace; and scope "observe" can only run read-only actions,
        else 403 forbidden_scope. /keys/* management always requires the wizard (no
        X-Key-Id). Constant-time compared. Query strings are NOT covered by the
        signature - destructive actions therefore ignore them; only read-only
        actions accept query params.
...
paths:
...
  /dayz/sources/vpp/{token}:
    post:
      summary: VPP event-feed ingress (Discord-shaped webhook sender)
      description: |
        Auth = the secret token in the URL (VPP_TOKEN), NOT HMAC - Discord-style
        senders can't set headers. Wrong token or feed disabled -> 404 (endpoint
        existence is not confirmed). Events are audited; opt-in Vpp.Rules can map
        event text to an action.
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                content: { type: string }
                embeds:
                  type: array
                  items:
                    type: object
                    properties:
                      title: { type: string }
                      description: { type: string }
            example: { content: 'Test event from Bruno', embeds: [] }
      responses:
        '204': { description: received and audited }
        '404': { description: wrong token or vpp disabled }